🚧 VQ05Q1 task: split context pipeline hotspots
