+++ .agentplane/tasks/202607092208-VQ05Q1/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 VQ05Q1 task: Split context pipeline hotspots for v0.6.22 [202607092208-VQ05Q1]
