+++ .agentplane/tasks/202607092208-VQ05Q1/pr/meta.json
@@ -1,0 +1,13 @@
+{
+  "base": "main",
+  "branch": "task/202607092208-VQ05Q1/split-context-pipeline-hotspots-for-v0-6-22",
+  "created_at": "2026-07-10T15:55:48.138Z",
+  "diffstat_sha256": "sha256:1e2ef810081e6c27e936f017e6edf27b2b6aaa22d0c902c99e8708bfbd2d33f3",
+  "last_verified_at": null,
+  "schema_version": 1,
+  "task_id": "202607092208-VQ05Q1",
+  "updated_at": "2026-07-10T15:55:48.138Z",
+  "verify": {
+    "status": "skipped"
+  }
+}
